requirements.txt
flask >= 2.0.0
pymongo >= 3.11.0
dnspython >= 2.1.0
python-dotenv >= 0.15.0
Flask-Cors==3.0.10
certifi